Class global vs Global
So, turns out there is a distinction to be made.  A variable declared
inside of a class is still referenced inside the method by self.x
reference.  However, if it is not declared as self.x it will be shared
among all instances (which is what we want in BigQuery case).  Still
need to refer to them as self.x inside methods using those variables
though.  Interesting.